refactor write_nmsg / write_nmsg_payload

+++ nmsg.h
@@ -108,10 +108,10 @@ extern nmsg_res		nmsg_io_add_pres(nmsg_io, nmsg_pres, nmsg_pbmod, void *);
 extern nmsg_res		nmsg_io_loop(nmsg_io);
 extern void		nmsg_io_destroy(nmsg_io *);
 extern void		nmsg_io_set_closed_fp(nmsg_io, nmsg_io_closed_fp);
-extern void		nmsg_io_set_count(nmsg_io, size_t);
+extern void		nmsg_io_set_count(nmsg_io, unsigned);
 extern void		nmsg_io_set_debug(nmsg_io, int);
 extern void		nmsg_io_set_endline(nmsg_io, const char *);
-extern void		nmsg_io_set_interval(nmsg_io, int);
+extern void		nmsg_io_set_interval(nmsg_io, unsigned);
 extern void		nmsg_io_set_output_mode(nmsg_io, nmsg_io_output_mode);
 
 /* nmsg_payload */
